:recycle: Add ProcessorOutput class to handle the optional loss return value from processor rollout
