+++ icenet_mp/types/complex_datatypes.py
@@ -1,11 +1,11 @@
 from collections.abc import Iterator, Mapping, Sequence
-from dataclasses import dataclass
-from typing import Any, Self
+from dataclasses import asdict, dataclass, field
+from typing import Any, Literal, Self, cast
 
-from omegaconf import DictConfig
+from omegaconf import DictConfig, OmegaConf
 from torch import Tensor
 
-from .typedefs import TensorNTCHW
+from .typedefs import DiffMode, DiffStrategy, TensorNTCHW
 
 
 class DataSpace:
@@ -46,19 +46,19 @@ class DataSpace:
 
 @dataclass
 class ModelStepOutput(Mapping[str, Tensor]):
-    """Output of a model step."""
+    """Output of a model step: prediction, target, and loss."""
 
     prediction: TensorNTCHW
-    target: TensorNTCHW | None = None
-    loss: Tensor | None = None
+    target: TensorNTCHW
+    loss: Tensor
 
     def __getitem__(self, key: str) -> Tensor:
         """Get a tensor by key."""
         if key == "prediction":
             return self.prediction
-        if key == "target" and self.target is not None:
+        if key == "target":
             return self.target
-        if key == "loss" and self.loss is not None:
+        if key == "loss":
             return self.loss
         msg = f"Key {key} not found in ModelStepOutput"
         raise KeyError(msg)
@@ -66,15 +66,102 @@ class ModelStepOutput(Mapping[str, Tensor]):
     def __iter__(self) -> Iterator[str]:
         """Iterate over the keys of ModelStepOutput."""
         yield "prediction"
-        if self.target is not None:
-            yield "target"
-        if self.loss is not None:
-            yield "loss"
+        yield "target"
+        yield "loss"
 
     def __len__(self) -> int:
         """Return ModelStepOutput length."""
-        return 1 + (self.target is not None) + (self.loss is not None)
+        return 3
 
     def copy(self) -> dict[str, Tensor]:
         """Return a plain dict copy, required by Lightning's manual optimization loop."""
         return dict(self)
+
+
+@dataclass
+class PlotSpec:
+    """Configure how sea-ice plots are rendered.
+
+    Attributes:
+        variable: Variable name shown in plots / used for routing.
+        title_groundtruth: Title above the ground-truth panel.
+        title_prediction: Title above the prediction panel.
+        title_difference: Title above the difference panel.
+        n_contour_levels: Number of contour levels per panel.
+        colourmap: colourmap used for GT/prediction panels.
+        include_difference: Whether to draw a difference panel.
+        diff_mode: Difference definition (e.g. "signed", "absolute", "smape").
+        diff_strategy: Strategy for animations (precompute, two-pass, per-frame).
+        selected_timestep: Slice index when a single timestep is needed.
+        vmin: Lower bound for GT/prediction colour scale (None = infer).
+        vmax: Upper bound for GT/prediction colour scale (None = infer).
+        colourbar_location: "vertical" or "horizontal".
+        colourbar_strategy: "shared" or "separate" colourbars.
+        outside_warn: Threshold for “values outside display range” warnings.
+        severe_outside: Severe threshold for clipping warnings.
+        include_shared_range_mismatch_check: If True, add magnitude mismatch nudges.
+
+    """
+
+    variable: str = "sea_ice_concentration"
+    title_groundtruth: str = "Ground Truth"
+    title_prediction: str = "Prediction"
+    title_difference: str = "Difference"
+
+    n_contour_levels: int = 51
+    colourmap: str = "viridis"
+
+    # Difference pane
+    include_difference: bool = True
+    diff_mode: DiffMode = "signed"
+    diff_strategy: DiffStrategy = "precompute"
+    selected_timestep: int = 0
+
+    # Colourscale ranges: defaults to [0,1]
+    vmin: float | None = 0.0
+    vmax: float | None = 1.0
+
+    # Colourbar layout
+    colourbar_location: Literal["vertical", "horizontal"] = "horizontal"
+    colourbar_strategy: Literal["shared", "separate"] = "shared"
+
+    # Range Check/warnings in badge
+    outside_warn: float = 0.05
+    severe_outside: float = 0.20
+    include_shared_range_mismatch_check: bool = True
+
+    # Optional metadata for titling
+    # hemisphere: "north" | "south" when known (used in titles)
+    hemisphere: Literal["north", "south"] | None = None
+    # metadata_subtitle: free-form text (e.g., "epochs=50; train=2010-2018")
+    metadata_subtitle: str | None = None
+
+    # Footer control
+    include_footer_metadata: bool = True
+
+    # Video settings
+    video_fps: int = 2
+    video_format: Literal["mp4", "gif"] = "mp4"
+
+    # Per-variable styles
+    per_variable_styles: dict[str, dict[str, str | float | bool]] = field(
+        default_factory=lambda: {
+            # Sea ice concentration
+            "sic-icenet:ice_conc": {"cmap": "Blues_r"},
+            "sic-ssmis:ice_conc": {"cmap": "Blues_r"},
+        }
+    )
+
+    def __add__(
+        self, other: "PlotSpec | DictConfig | dict [str, Any] | None"
+    ) -> "PlotSpec":
+        """Combine two PlotSpec instances or a PlotSpec with a dictionary."""
+        if other is None:
+            return self
+        if isinstance(other, PlotSpec):
+            dict_other = asdict(other)
+        elif isinstance(other, DictConfig):
+            dict_other = cast("dict[str, Any]", OmegaConf.to_container(other))
+        else:
+            dict_other = dict(other)
+        return PlotSpec(**(asdict(self) | dict_other))
